requirements.txt
elasticsearch==8.11.0
psycopg2-binary==2.9.9
Faker==20.1.0
requests==2.31.0
